Fixed notifications for latest FF 48.0a2
There is a difference in the notifications API between Firefox and Chrome. Firefox hasn't ever supported buttons inside notifications like Chrome, but wasn't throwing an exception. But instead of implementing this missing feature they're throwing an exception now. This of course breaks the way this extension creates notifications. I've created an issue on BugZilla (http://tinyurl.com/jxjgg5o), but the developers think that this is a correct behavior.

Hopefully Firefox will fix the behavior to act like Chrome soon.